+++ hc/accounts/models.py
@@ -182,13 +182,6 @@ class Profile(models.Model):
         }
         emails.transfer_request(self.user.email, ctx)
 
-    def send_sms_limit_notice(self, transport: str) -> None:
-        ctx = {"transport": transport, "limit": self.sms_limit}
-        if self.sms_limit != 500 and settings.USE_PAYMENTS:
-            ctx["url"] = absolute_reverse("hc-pricing")
-
-        emails.sms_limit(self.user.email, ctx)
-
     def projects(self) -> QuerySet[Project]:
         """Return a queryset of all projects we have access to."""