+++ hc/integrations/whatsapp/tests/test_notify.py
@@ -123,7 +123,15 @@ class NotifyWhatsAppTestCase(BaseTestCase):
 
         email = mail.outbox[0]
         self.assertEqual(email.to[0], "alice@example.org")
+        self.assertEqual(email.to[1], "bob@example.org")
         self.assertEqual(email.subject, "Monthly WhatsApp Limit Reached")
+
+        # Account's owner should be mentioned in the message body
+        self.assertEmailContains("alice@example.org")
+
+        # The alert content itself should be in the message body
+        self.assertEmailContainsText("""The check "Foo" is DOWN""")
+        self.assertEmailContainsHtml("""The check &quot;Foo&quot; is DOWN""")
 
     @patch("hc.api.transports.curl.request", autospec=True)
     def test_it_does_not_escape_special_characters(self, mock_post: Mock) -> None: