+++ hc/integrations/whatsapp/transport.py
@@ -5,11 +5,12 @@ import logging
 from typing import NoReturn
 
 from django.conf import settings
+from pydantic import BaseModel, ValidationError
+
 from hc.accounts.models import Profile
 from hc.api.models import Flip, Notification
 from hc.api.transports import HttpTransport, TransportError
 from hc.lib import curl
-from pydantic import BaseModel, ValidationError
 
 logger = logging.getLogger(__name__)
 
@@ -55,9 +56,20 @@ class WhatsApp(HttpTransport):
             if not getattr(settings, key):
                 raise TransportError("WhatsApp notifications are not enabled")
 
-        profile = Profile.objects.for_user(self.channel.project.owner)
+        profile = self.channel.project.owner_profile
         if not profile.authorize_sms():
-            profile.send_sms_limit_notice("WhatsApp")
+            # We don't know the precise message content (the message templates
+            # are defined and live in Twilio). Use the SMS template as a
+            # "close enough" substitute:
+            text = self.tmpl(
+                "sms_message.html",
+                flip=flip,
+                check=flip.owner,
+                status=flip.new_status,
+                site_name=settings.SITE_NAME,
+            )
+
+            self.channel.send_sms_limit_notice("WhatsApp", text)
             raise TransportError("Monthly message limit exceeded")
 
         url = self.URL % settings.TWILIO_ACCOUNT